Adds means to list file share contents
